# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/blog-api/category/v1/list": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "list Category",
                "description": "list Category",
                "operationId": "list",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skip",
                                    "take",
                                    "issue_type"
                                ],
                                "properties": {
                                    "skip": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0"
                                    },
                                    "take": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "10"
                                    },
                                    "issue_type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "article"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {},
                                            "example": [
                                                {
                                                    "name": "Post",
                                                    "title": "title Post",
                                                    "slug": "slug Post",
                                                    "subject": "subject Post",
                                                    "pic_small": "pic_small Post",
                                                    "updated_at": "2023-02-19T07:39:12.000000Z",
                                                    "created_at": "2023-02-19T07:39:12.000000Z",
                                                    "id": "1"
                                                },
                                                {
                                                    "name": "Post",
                                                    "title": "title Post",
                                                    "slug": "slug Post",
                                                    "subject": "subject Post",
                                                    "pic_small": "pic_small Post",
                                                    "updated_at": "2023-02-19T07:39:12.000000Z",
                                                    "created_at": "2023-02-19T07:39:12.000000Z",
                                                    "id": "2"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error  edit"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/blog-api/category/v1/store": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Store category in DB",
                "description": "Store category in DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "pic_small": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "pic_large": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolaen",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "Create new Category"
                                        },
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {},
                                            "example": {
                                                "name": "category name",
                                                "title": "category post",
                                                "status": "True",
                                                "slug": "slug category",
                                                "description": "description post",
                                                "meta_description": "meta_description post",
                                                "meta_keywords": "meta_keywords post",
                                                "meta_language": "meta_language post",
                                                "tweet_text": "tweet_text post",
                                                "issue_type": "type",
                                                "menu_order": 1,
                                                "priority": "133",
                                                "pic_small": "pic_small post",
                                                "pic_large": "pic_large post",
                                                "menu_status": true,
                                                "visible_index_status": true,
                                                "parent_id": "1",
                                                "updated_at": "2023-02-19T07:39:12.000000Z",
                                                "created_at": "2023-02-19T07:39:12.000000Z",
                                                "id": "1"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error  edit"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/blog-api/category/v1/soft-delete": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Soft-delete category in DB",
                "description": "Soft-delete category in DB",
                "operationId": "soft-delete-category",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error soft delete"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}